feat: Add delete and edit photo functionality

+++ app/Http/Controllers/PhotoController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\StorePhotoRequest;
+use App\Http\Requests\UpdatePhotoRequest;
 use App\Models\Photo;
 
 class PhotoController extends Controller
@@ -37,8 +38,34 @@ class PhotoController extends Controller
         return redirect()->route('photos.index');
     }
 
+    public function edit(Photo $photo)
+    {
+        $this->authorize('update', $photo);
+
+        return view('photos.edit', [
+            'photo' => $photo,
+        ]);
+    }
+
+    public function update(UpdatePhotoRequest $request, Photo $photo)
+    {
+        $this->authorize('update', $photo);
+
+        $validated = $request->validated();
+
+        if ($request->hasFile('photo')) {
+            $validated['src'] = $request->file('photo')->store('photos');
+        }
+
+        $photo->update($validated);
+
+        return redirect()->route('photos.show', $photo);
+    }
+
     public function destroy(Photo $photo)
     {
+        $this->authorize('delete', $photo);
+
         $photo->delete();
 
         return redirect()->route('photos.index');
